feat: integração com backend para componente Grupo

+++ BackEnd/src/main/java/com/spring/boot/controlefinanceiro/service/GrupoService.java
@@ -43,6 +43,13 @@ public class GrupoService {
     }
 
     public void delete(Long id) {
-        this.repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new NotFoundException("Grupo não encontrada");
+        }
+        try {
+            this.repository.deleteById(id);
+        } catch (Exception e) {
+            throw new RegraNegocioException("Erro ao tentar excluir o grupo. Verifique se há registros associados.");
+        }
     }
 }
